Cut out the "Web" from project name
