+++ src/Legacy/Legacy.Monolith.Api/Controllers/RestrictionsController.cs
@@ -2,15 +2,15 @@ using Legacy.Monolith.DbContexts;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
-namespace Legacy.Monolith.WebApi.Controllers;
+namespace Legacy.Monolith.Api.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class BrandsController : ControllerBase
+public class RestrictionsController : ControllerBase
 {
     private readonly CatalogDbContext _dbContext;
 
-    public BrandsController(CatalogDbContext dbContext)
+    public RestrictionsController(CatalogDbContext dbContext)
     {
         _dbContext = dbContext;
     }
@@ -18,14 +18,14 @@ public class BrandsController : ControllerBase
     [HttpGet("")]
     public async Task<IActionResult> GetAll()
     {
-        var result = await _dbContext.Brands.ToListAsync();
+        var result = await _dbContext.Restrictions.ToListAsync();
         return Ok(result);
     }
 
     [HttpGet("{id:int}")]
     public async Task<IActionResult> GetById(int id)
     {
-        var result = await _dbContext.Brands.FirstOrDefaultAsync(x => x.Id == id);
+        var result = await _dbContext.Restrictions.FirstOrDefaultAsync(x => x.Id == id);
         return Ok(result);
     }
 }
