+++ src/Legacy/Legacy.Monolith.Api/Legacy.Monolith.Api.csproj
