+++ src/Legacy/Legacy.Monolith.Api/Properties/launchSettings.json
