+++ src/Legacy/Legacy.Monolith.Api/appsettings.Development.json
